move notice to domain
