+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/CannotDownloadArchiveFromNetworkNotice.java
@@ -14,31 +14,34 @@
  * limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.usecase.notice.error;
+package org.mobilitydata.gtfsvalidator.domain.entity.notice.error;
 
-import org.mobilitydata.gtfsvalidator.usecase.notice.base.ErrorNotice;
-import org.mobilitydata.gtfsvalidator.usecase.port.ValidationResultRepository;
+import org.mobilitydata.gtfsvalidator.domain.entity.NoticeExporter;
+import org.mobilitydata.gtfsvalidator.domain.entity.notice.base.ErrorNotice;
 
 import java.io.IOException;
+import java.net.URL;
 
-public class MissingHeaderNotice extends ErrorNotice {
-    private String missingHeaderName;
+public class CannotDownloadArchiveFromNetworkNotice extends ErrorNotice {
+    private URL url;
 
-    public MissingHeaderNotice(final String filename, final String missingHeaderName) {
-        super(filename, E_001,
-                "Missing required header",
-                "File " + filename + " is missing required header: " + missingHeaderName,
+    public CannotDownloadArchiveFromNetworkNotice(final URL url) {
+        super(String.valueOf(url),
+                E_007,
+                "Downloading error",
+                "An error occurred while trying to download archive from network: " + url,
                 null);
-        this.missingHeaderName = missingHeaderName;
+        this.url = url;
     }
+
+    public URL getUrl() {
+        return url;
+    }
+
 
     @Override
     public void export(ValidationResultRepository.NoticeExporter exporter)
             throws IOException {
         exporter.export(this);
-    }
-
-    public String getMissingHeaderName() {
-        return missingHeaderName;
     }
 }
