+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/CannotParseFloatNotice.java
@@ -14,23 +14,23 @@
  * limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.usecase.notice.error;
+package org.mobilitydata.gtfsvalidator.domain.entity.notice.error;
 
-import org.mobilitydata.gtfsvalidator.usecase.notice.base.ErrorNotice;
-import org.mobilitydata.gtfsvalidator.usecase.port.ValidationResultRepository;
+import org.mobilitydata.gtfsvalidator.domain.entity.NoticeExporter;
+import org.mobilitydata.gtfsvalidator.domain.entity.notice.base.ErrorNotice;
 
 import java.io.IOException;
 
-public class CannotParseIntegerNotice extends ErrorNotice {
+public class CannotParseFloatNotice extends ErrorNotice {
     private int lineNumber;
     private String rawValue;
     private String fieldName;
 
-    public CannotParseIntegerNotice(String filename, String fieldName, int lineNumber, String rawValue) {
-        super(filename, E_005,
-                "Invalid integer value",
+    public CannotParseFloatNotice(String filename, String fieldName, int lineNumber, String rawValue) {
+        super(filename, E_006,
+                "Invalid float value",
                 "Value: '" + rawValue + "' of field: " + fieldName
-                        + " with type integer can't be parsed in file: " + filename + " at row: " + lineNumber,
+                        + " with type float can't be parsed in file: " + filename + " at row: " + lineNumber,
                 null);
         this.rawValue = rawValue;
         this.fieldName = fieldName;
