+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/InvalidUrlNotice.java
@@ -14,25 +14,24 @@
  * limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.usecase.notice.error;
+package org.mobilitydata.gtfsvalidator.domain.entity.notice.error;
 
-import org.mobilitydata.gtfsvalidator.usecase.notice.base.ErrorNotice;
-import org.mobilitydata.gtfsvalidator.usecase.port.ValidationResultRepository;
+import org.mobilitydata.gtfsvalidator.domain.entity.NoticeExporter;
+import org.mobilitydata.gtfsvalidator.domain.entity.notice.base.ErrorNotice;
 
 import java.io.IOException;
 
-public class InvalidTimezoneNotice extends ErrorNotice {
+public class InvalidUrlNotice extends ErrorNotice {
+    private String urlValue;
     private String fieldName;
-    private String timezoneValue;
 
-    public InvalidTimezoneNotice(String filename, String fieldName, String entityId, String timezoneValue) {
-        super(filename, E_013,
-                "Invalid timezone",
-                "Invalid timezone:" + timezoneValue + " in field:" + fieldName
-                        + " for entity with id:" + entityId,
+    public InvalidUrlNotice(String filename, String fieldName, String entityId, String urlValue) {
+        super(filename, E_012,
+                "Invalid url",
+                "Invalid url:" + urlValue + " in field:" + fieldName + " for entity with id:" + entityId,
                 entityId);
         this.fieldName = fieldName;
-        this.timezoneValue = timezoneValue;
+        this.urlValue = urlValue;
     }
 
     @Override
@@ -41,15 +40,15 @@ public class InvalidTimezoneNotice extends ErrorNotice {
         exporter.export(this);
     }
 
+    public String getUrlValue() {
+        return urlValue;
+    }
+
     public String getEntityId() {
         return entityId;
     }
 
     public String getFieldName() {
         return fieldName;
-    }
-
-    public String getTimezoneValue() {
-        return timezoneValue;
     }
 }
