+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/MissingHeaderNotice.java
@@ -14,25 +14,31 @@
  * limitations under the License.
  */
 
-package org.mobilitydata.gtfsvalidator.usecase.notice.error;
+package org.mobilitydata.gtfsvalidator.domain.entity.notice.error;
 
-import org.mobilitydata.gtfsvalidator.usecase.notice.base.ErrorNotice;
-import org.mobilitydata.gtfsvalidator.usecase.port.ValidationResultRepository;
+import org.mobilitydata.gtfsvalidator.domain.entity.NoticeExporter;
+import org.mobilitydata.gtfsvalidator.domain.entity.notice.base.ErrorNotice;
 
 import java.io.IOException;
 
-public class CannotConstructDataProviderNotice extends ErrorNotice {
+public class MissingHeaderNotice extends ErrorNotice {
+    private String missingHeaderName;
 
-    public CannotConstructDataProviderNotice(String filename) {
-        super(filename, E_002,
-                "Data provider error",
-                "An error occurred while trying to access raw data for file: " + filename,
+    public MissingHeaderNotice(final String filename, final String missingHeaderName) {
+        super(filename, E_001,
+                "Missing required header",
+                "File " + filename + " is missing required header: " + missingHeaderName,
                 null);
+        this.missingHeaderName = missingHeaderName;
     }
 
     @Override
     public void export(ValidationResultRepository.NoticeExporter exporter)
             throws IOException {
         exporter.export(this);
+    }
+
+    public String getMissingHeaderName() {
+        return missingHeaderName;
     }
 }
