+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/notice/error/MissingRequiredValueNotice.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2020. MobilityData IO.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.mobilitydata.gtfsvalidator.domain.entity.notice.error;
+
+import org.mobilitydata.gtfsvalidator.domain.entity.NoticeExporter;
+import org.mobilitydata.gtfsvalidator.domain.entity.notice.base.ErrorNotice;
+
+import java.io.IOException;
+
+public class MissingRequiredValueNotice extends ErrorNotice {
+    private String fieldName;
+
+    public MissingRequiredValueNotice(String filename, String fieldName, String entityId) {
+        super(filename, E_015,
+                "Missing required value",
+                "Missing value for field:" + fieldName
+                        + " marked as required in entity with id:" + entityId,
+                entityId);
+        this.fieldName = fieldName;
+    }
+
+    @Override
+    public void export(ValidationResultRepository.NoticeExporter exporter)
+            throws IOException {
+        exporter.export(this);
+    }
+
+    public String getFieldName() {
+        return fieldName;
+    }
+
+}
